more splitting up stat logic and letting us know what is the team size as its created
